Applies the Multimap near cache optimization from vertx-hazelcast
Since we now have near cache, the infinispan cache no longer needs to be replicated.

My testing with the request/reply scenario show that CPU usage is similar to vertx-hazelcast after the change.
Before, one my CPU was around 100% at 500msg/sec, because we had to iterate through the set of mappings.