Configuration process update
The JGroups file has been removed from the project.
This is possible since #16 because the lock and counter protocols are added to a forked channel.
Consequently, we can use the default-configs/default-jgroups-tcp.xml file from the ISPN project.
This simplifies ISPN version changes as we don't have to check for upstream JGroups config changes.

Also, using the cluster manager, I realized that you often need to update the JGroups config, not the ISPN config.
So now you can configure it with infinispan.xml and/or jgroups.xml files on the classpath.
Alternatively, you can set the vertx.infinispan.config and/or vertx.jgroups.config sysprops.

Eventually, the Openshift3 setup has been greatly simplified.
Thanks to the vertx.jgroups.config sysprop, all you need is the
infinispan-cloud dependency. infinispan-cloud brings jgroups-kubernetes
(KUBE_PING) as well as a preconfigured Kubernetes/Openshift3 stack.