Upgrade to ISPN  (#148)
* Dependency upgrades for ISPN 15

Moved dependency convergence constraints to the dependencyManagement section in the parent POM.

Also:
- upgraded to Caffeine 3.2.1 to match vertx-web version
- upgraded testcontainers

Signed-off-by: Thomas Segismont <tsegismont@gmail.com>

* DefaultTemplate is deprecated for removal in IPSN 15

Signed-off-by: Thomas Segismont <tsegismont@gmail.com>

* EmbeddedCacheManager no longer exposes the transport

As advised by ISPN team, use the global component registry to retrieve it.

Signed-off-by: Thomas Segismont <tsegismont@gmail.com>

* XML schema updates

Signed-off-by: Thomas Segismont <tsegismont@gmail.com>

* ISPN 15 requires JDK 17+

Configure Maven toolchains plugin in a profile that is enabled when the Java version is less than 17.

In CI, make SDK 17 available for toolchains.

Signed-off-by: Thomas Segismont <tsegismont@gmail.com>

* Update JGroups test config

Align with ISPN default TCP stack

Signed-off-by: Thomas Segismont <tsegismont@gmail.com>

---------

Signed-off-by: Thomas Segismont <tsegismont@gmail.com>